pyproject.toml
[project]
name = "nlp100"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121" },
]
torchvision = [
    { index = "pytorch-cu121" },
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
